Split articulation ordering resolvers
Move backend convention discovery helpers out of ordering.py into ordering_resolvers.py so ordering.py stays focused on the public ordering data model and map construction.
